Slf4j has the concept of markers, which allows users to mark log messages.  Logback currently implements this idea, but the gelf converter was ignoring the markers in the ILoggingEvent object that it had to convert.
I modified this code to have a flag called useMarker that tells the converter whether or not it should include markers in the GELF message.  This flag was added both to the GelfAppender and GelfLayout so that it could be passed through to the underlying GelfConverter.  I followed the pattern established by similar flags such as useLoggerName.